Add prefetch of instructions to be executed.
In order to avoid cache misses in instruction loading code,
explicitly prefetch instructions that are likely to be called
soon. For simplicity, do the lookahead in the queue of the
current warrior, and don't handle queue wrapping.

The queue buffer should be cleared before usage to avoid
prefetching random garbage.