Move the current value retrieval to the common arg handler.
Update operations need to know the contents of an instruction
after all post-increments. This need is common enough to be
implemented directly in the argument load code.

Since CUR_CMD_XMM isn't used after CUR_B_XMM is computed, they
are mapped to the same register. It also saves a move in the
most trivial case.