Support running only one warrior.
The scheduler scheme assumes that there are at least two threads
running, which is always correct in the multiple warrior case.
If there is only one warrior, it is necessary to inject a stub
thread.

The fake instruction executed by the thread is written immediately
beyond the end of the core, where it is inaccessible to ordinary
code due to wrapping, but visible to the scheduler.

The stub thread suicides if the warrior creates multiple threads
of its own, and is automatically resurrected when they die.